# docs/api/API-contract.yaml
openapi: 3.0.3
info:
  title: Easyfamily API Contract
  version: 0.1.0
  description: |
    Contract-first API definition for Easyfamily backend.
    All successful and business-failed responses follow ApiResponse envelope.
    Public error codes are client-facing only and never expose internal exception codes.
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/reports/dau:
    get:
      tags: [Admin]
      summary: DAU report
      responses:
        '200':
          description: Api response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDauReport'
        '401':
          description: Missing/invalid access token
  /api/v1/admin/reports/feature-hot:
    get:
      tags: [Admin]
      summary: Feature hot report
      responses:
        '200':
          description: Api response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFeatureHotReport'
        '401':
          description: Missing/invalid access token
  /api/v1/admin/reports/query-overview:
    get:
      tags: [Admin]
      summary: Query overview report
      responses:
        '200':
          description: Api response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQueryOverview'
        '401':
          description: Missing/invalid access token
  /api/v1/admin/query-settings:
    get:
      tags: [Admin]
      summary: Get runtime query settings
      responses:
        '200':
          description: Api response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQuerySettings'
        '401':
          description: Missing/invalid access token
  /api/v1/admin/quota:
    patch:
      tags: [Admin]
      summary: Update runtime query quota/settings
      parameters:
        - in: query
          name: dailyQuotaPerUser
          schema:
            type: integer
            minimum: 1
        - in: query
          name: dailyQuotaPerPhone
          schema:
            type: integer
            minimum: 1
        - in: query
          name: dailyQuotaPerIp
          schema:
            type: integer
            minimum: 1
        - in: query
          name: preferRedisCache
          schema:
            type: boolean
        - in: query
          name: providerKey
          schema:
            type: string
        - in: query
          name: providerTimeoutMs
          schema:
            type: integer
            minimum: 100
        - in: query
          name: providerRetryTimes
          schema:
            type: integer
            minimum: 0
        - in: query
          name: providerCircuitFailureThreshold
          schema:
            type: integer
            minimum: 1
        - in: query
          name: providerCircuitOpenSeconds
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Api response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQuerySettings'
        '401':
          description: Missing/invalid access token
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseDauReport:
      type: object
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data:
          type: array
          items:
            type: object
            properties:
              date: { type: string, example: "2026-03-07" }
              dau: { type: integer, example: 123 }
      required: [code, message, data]
    ApiResponseFeatureHotReport:
      type: object
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data:
          type: array
          items:
            type: object
            properties:
              feature: { type: string, example: phone-binding-query }
              pv: { type: integer, example: 321 }
      required: [code, message, data]
    ApiResponseQueryOverview:
      type: object
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data:
          $ref: '#/components/schemas/QueryOverview'
      required: [code, message, data]
    ApiResponseQuerySettings:
      type: object
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data:
          $ref: '#/components/schemas/QuerySettings'
      required: [code, message, data]
...
    QueryOverview:
      type: object
      properties:
        totalQueryCount: { type: integer, example: 10 }
        cacheHitCount: { type: integer, example: 7 }
        cacheHitRate: { type: number, format: double, example: 0.7 }
        dailyQuotaPerUser: { type: integer, example: 3 }
        dailyQuotaPerPhone: { type: integer, example: 30 }
        dailyQuotaPerIp: { type: integer, example: 100 }
        preferRedisCache: { type: boolean, example: true }
        providerKey: { type: string, example: aliyun-market }
        providerTimeoutMs: { type: integer, example: 1500 }
        providerRetryTimes: { type: integer, example: 1 }
        providerCircuitFailureThreshold: { type: integer, example: 5 }
        providerCircuitOpenSeconds: { type: integer, example: 30 }
    QuerySettings:
      type: object
      properties:
        dailyQuotaPerUser: { type: integer, example: 3 }
        dailyQuotaPerPhone: { type: integer, example: 30 }
        dailyQuotaPerIp: { type: integer, example: 100 }
        preferRedisCache: { type: boolean, example: true }
        providerKey: { type: string, example: aliyun-market }
        providerTimeoutMs: { type: integer, example: 1500 }
        providerRetryTimes: { type: integer, example: 1 }
        providerCircuitFailureThreshold: { type: integer, example: 5 }
        providerCircuitOpenSeconds: { type: integer, example: 30 }